fix: d-day 타입 수정, 자동 전송 기능 수정

+++ src/main/java/com/myprecious/moneyglove/domain/board/dto/response/BoardDDayResponse.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BoardDDayResponse {
-    private Long DDay;
+    private Integer DDay;
 
     public BoardDDayResponse(BoardEntity boardEntity){
         this.DDay = boardEntity.getDDay();
